@@ -6,8 +6,8 @@
         "@reduxjs/toolkit": "^1.7.2",
         "axios": "^0.21.4",
         "connected-react-router": "^6.9.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-redux": "^7.2.6",
         "react-router-dom": "^5.2.0",
         "redux-saga": "^1.1.3"
